pyproject.toml
[tool.poetry]
name = "chat-fictional-character"
version = "0.1.0"
description = "Chat with fictional character."
authors = ["Simone Appella"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
numpy = "^1.26.0"
flask = "^3.0.3"
faiss-cpu = "^1.9.0"
openai = "^1.53.0"
pyarrow = "^18.0.0"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.5"
langchain-community = "^0.3.5"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
setuptools = "^75.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
